1.解决分享回来闪烁问题 2.优化重构微信module
