+++ android/build.gradle
@@ -1,27 +1,30 @@
 apply plugin: 'com.android.library'
 
 android {
-  compileSdkVersion 23
-  buildToolsVersion "25.0.0"
-  defaultConfig {
-    minSdkVersion 16
-    targetSdkVersion 22
-    versionCode 1
-    versionName "1.0"
-    ndk {
-      abiFilters "armeabi-v7a", "x86"
+    compileSdkVersion 23
+    buildToolsVersion "25.0.0"
+    defaultConfig {
+        minSdkVersion 16
+        targetSdkVersion 22
+        versionCode 1
+        versionName "1.0"
+        ndk {
+            abiFilters "armeabi-v7a", "x86"
+        }
     }
-  }
 }
 
 allprojects {
-  repositories {
-    jcenter()
-    maven { url "$projectDir/../../react-native/android" }
-  }
+    repositories {
+        jcenter()
+        maven { url "$projectDir/../../react-native/android" }
+    }
 }
 
 dependencies {
-  compile 'com.facebook.react:react-native:+'
-  compile files('libs/libammsdk.jar')
+    compile 'com.facebook.react:react-native:+'
+    compile files('libs/libammsdk.jar')
+    provided 'io.reactivex:rxjava:1.3.0'
+    provided 'io.reactivex:rxandroid:1.2.1'
+    compile 'me.comori.tools:databus:1.0.1'
 }
