+++ android/src/main/AndroidManifest.xml
@@ -5,4 +5,20 @@
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
     <uses-permission android:name="android.permission.READ_PHONE_STATE"/>
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
+
+    <application>
+
+        <activity android:name="com.theweflex.react.WXEntryActivity"
+            android:exported="true"
+            android:theme="@android:style/Theme.Translucent.NoTitleBar"
+            android:launchMode="singleTask"
+            />
+        
+        <activity-alias
+            android:name="${applicationId}.wxapi.WXEntryActivity"
+            android:targetActivity="com.theweflex.react.WXEntryActivity"
+            android:exported="true"
+            />
+
+    </application>
 </manifest>
