+++ android/src/main/java/com/theweflex/react/WeChatModule.java
@@ -1,29 +1,6 @@
 package com.theweflex.react;
 
-import android.content.Context;
-import android.content.Intent;
-import android.graphics.Bitmap;
-import android.net.Uri;
-import android.support.annotation.Nullable;
-import android.webkit.MimeTypeMap;
-
-import com.facebook.common.executors.UiThreadImmediateExecutorService;
-import com.facebook.common.internal.Closeables;
-import com.facebook.common.internal.Preconditions;
-import com.facebook.common.memory.PooledByteBuffer;
-import com.facebook.common.memory.PooledByteBufferInputStream;
-import com.facebook.common.references.CloseableReference;
-import com.facebook.common.util.UriUtil;
-import com.facebook.datasource.BaseDataSubscriber;
-import com.facebook.datasource.DataSource;
-import com.facebook.datasource.DataSubscriber;
-import com.facebook.drawee.backends.pipeline.Fresco;
-import com.facebook.imagepipeline.common.ResizeOptions;
-import com.facebook.imagepipeline.core.ImagePipeline;
-import com.facebook.imagepipeline.datasource.BaseBitmapDataSubscriber;
-import com.facebook.imagepipeline.image.CloseableImage;
-import com.facebook.imagepipeline.request.ImageRequest;
-import com.facebook.imagepipeline.request.ImageRequestBuilder;
+import com.comori.tools.databus.DataBus;
 import com.facebook.react.bridge.Arguments;
 import com.facebook.react.bridge.Callback;
 import com.facebook.react.bridge.ReactApplicationContext;
@@ -32,51 +9,26 @@ import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.modules.core.DeviceEventManagerModule;
-import com.facebook.react.modules.network.OkHttpClientProvider;
-import com.tencent.mm.sdk.modelbase.BaseReq;
 import com.tencent.mm.sdk.modelbase.BaseResp;
 import com.tencent.mm.sdk.modelmsg.SendAuth;
 import com.tencent.mm.sdk.modelmsg.SendMessageToWX;
-import com.tencent.mm.sdk.modelmsg.WXEmojiObject;
-import com.tencent.mm.sdk.modelmsg.WXFileObject;
-import com.tencent.mm.sdk.modelmsg.WXImageObject;
-import com.tencent.mm.sdk.modelmsg.WXMediaMessage;
-import com.tencent.mm.sdk.modelmsg.WXMusicObject;
-import com.tencent.mm.sdk.modelmsg.WXTextObject;
-import com.tencent.mm.sdk.modelmsg.WXVideoObject;
-import com.tencent.mm.sdk.modelmsg.WXWebpageObject;
-import com.tencent.mm.sdk.modelpay.PayReq;
 import com.tencent.mm.sdk.modelpay.PayResp;
-import com.tencent.mm.sdk.openapi.IWXAPI;
-import com.tencent.mm.sdk.openapi.IWXAPIEventHandler;
-import com.tencent.mm.sdk.openapi.WXAPIFactory;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.UUID;
-
-import okhttp3.Call;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
+import com.theweflex.react.exception.InvalidArgumentException;
+import com.theweflex.react.exception.InvokeException;
+import com.theweflex.react.exception.NotRegisterException;
 
 /**
  * Created by tdzl2_000 on 2015-10-10.
  */
-public class WeChatModule extends ReactContextBaseJavaModule implements IWXAPIEventHandler {
-    private String appId;
+public class WeChatModule extends ReactContextBaseJavaModule {
 
-    private IWXAPI api = null;
-    private final static String NOT_REGISTERED = "registerApp required.";
-    private final static String INVOKE_FAILED = "WeChat API invoke returns false.";
-    private final static String INVALID_ARGUMENT = "invalid argument.";
-    private HashMap<String, String> extensionMap = new HashMap<>();
+    public static final String KEY_WX_RESULT = "WX_RESULT";
+
+    private WXPresenter mWxPresenter;
 
     public WeChatModule(ReactApplicationContext context) {
         super(context);
+        mWxPresenter = new WXPresenter(context);
     }
 
     @Override
@@ -94,591 +46,140 @@ public class WeChatModule extends ReactContextBaseJavaModule implements IWXAPIEv
         return true;
     }
 
-    private static ArrayList<WeChatModule> modules = new ArrayList<>();
-
     @Override
     public void initialize() {
         super.initialize();
-        modules.add(this);
+        DataBus.get().with(KEY_WX_RESULT, BaseResp.class)
+                .observe(onWxResultOB);
     }
 
     @Override
     public void onCatalystInstanceDestroy() {
         super.onCatalystInstanceDestroy();
-        if (api != null) {
-            api = null;
+        if (mWxPresenter != null) {
+            mWxPresenter.destroy();
         }
-        modules.remove(this);
-    }
-
-    public static void handleIntent(Intent intent) {
-        for (WeChatModule mod : modules) {
-            if (mod.api != null) {
-                mod.api.handleIntent(intent, mod);
-            }
-        }
+        DataBus.get().remove(KEY_WX_RESULT);
     }
 
     @ReactMethod
     public void registerApp(String appid, Callback callback) {
-        this.appId = appid;
-        api = WXAPIFactory.createWXAPI(this.getReactApplicationContext().getBaseContext(), appid, true);
-        callback.invoke(null, api.registerApp(appid));
+        mWxPresenter.register(appid);
     }
 
     @ReactMethod
     public void isWXAppInstalled(Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
+        try {
+            callback.invoke(null, mWxPresenter.isWXAppInstalled());
+        } catch (NotRegisterException e) {
+            callback.invoke(e.getMessage());
         }
-        callback.invoke(null, api.isWXAppInstalled());
     }
 
     @ReactMethod
     public void isWXAppSupportApi(Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
+        try {
+            callback.invoke(null, mWxPresenter.isWXAppSupportApi());
+        } catch (NotRegisterException e) {
+            callback.invoke(e.getMessage());
         }
-        callback.invoke(null, api.isWXAppSupportAPI());
     }
 
     @ReactMethod
     public void getApiVersion(Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
+        try {
+            callback.invoke(null, mWxPresenter.getApiVersion());
+        } catch (NotRegisterException e) {
+            callback.invoke(e.getMessage());
         }
-        callback.invoke(null, api.getWXAppSupportAPI());
     }
 
     @ReactMethod
     public void openWXApp(Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
+        try {
+            callback.invoke(null, mWxPresenter.openWXApp());
+        } catch (NotRegisterException e) {
+            callback.invoke(e.getMessage());
         }
-        callback.invoke(null, api.openWXApp());
     }
 
     @ReactMethod
     public void sendAuthRequest(String scope, String state, Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
+        try {
+            callback.invoke(null, mWxPresenter.sendAuthRequest(scope, state));
+        } catch (NotRegisterException e) {
+            callback.invoke(e.getMessage());
         }
-        SendAuth.Req req = new SendAuth.Req();
-        req.scope = scope;
-        req.state = state;
-        callback.invoke(null, api.sendReq(req));
+    }
+
+    public void share(int scene, ReadableMap data, final Callback callback) {
+        try {
+            mWxPresenter.rnShare(scene, data, new WXPresenter.ShareCallback() {
+                @Override
+                public void onShareCompleted(boolean success) {
+                    callback.invoke(null, success);
+                }
+            });
+        } catch (InvalidArgumentException | NotRegisterException e) {
+            callback.invoke(e.getMessage());
+        }
     }
 
     @ReactMethod
-    public void shareToTimeline(ReadableMap data, Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
-        }
-        _share(SendMessageToWX.Req.WXSceneTimeline, data, callback);
+    public void shareToTimeline(ReadableMap data, final Callback callback) {
+        share(SendMessageToWX.Req.WXSceneTimeline, data, callback);
     }
 
     @ReactMethod
     public void shareToSession(ReadableMap data, Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
-        }
-        _share(SendMessageToWX.Req.WXSceneSession, data, callback);
+        share(SendMessageToWX.Req.WXSceneSession, data, callback);
     }
 
     @ReactMethod
     public void shareToFavorite(ReadableMap data, Callback callback) {
-        if (api == null) {
-            callback.invoke(NOT_REGISTERED);
-            return;
-        }
-        _share(SendMessageToWX.Req.WXSceneFavorite, data, callback);
+        share(SendMessageToWX.Req.WXSceneFavorite, data, callback);
     }
 
     @ReactMethod
     public void pay(ReadableMap data, Callback callback) {
-        PayReq payReq = new PayReq();
-        if (data.hasKey("partnerId")) {
-            payReq.partnerId = data.getString("partnerId");
-        }
-        if (data.hasKey("prepayId")) {
-            payReq.prepayId = data.getString("prepayId");
-        }
-        if (data.hasKey("nonceStr")) {
-            payReq.nonceStr = data.getString("nonceStr");
-        }
-        if (data.hasKey("timeStamp")) {
-            payReq.timeStamp = data.getString("timeStamp");
-        }
-        if (data.hasKey("sign")) {
-            payReq.sign = data.getString("sign");
-        }
-        if (data.hasKey("package")) {
-            payReq.packageValue = data.getString("package");
-        }
-        if (data.hasKey("extData")) {
-            payReq.extData = data.getString("extData");
-        }
-        payReq.appId = appId;
-        callback.invoke(api.sendReq(payReq) ? null : INVOKE_FAILED);
-    }
-
-    private void _share(final int scene, final ReadableMap data, final Callback callback) {
-        Uri thumbUri = null;
-        if (data.hasKey("thumbImage")) {
-            String thumbImage = data.getString("thumbImage");
-            try {
-                thumbUri = Uri.parse(thumbImage);
-                // Verify scheme is set, so that relative uri (used by static resources) are not handled.
-                if (thumbUri.getScheme() == null) {
-                    thumbUri = getResourceDrawableUri(getReactApplicationContext(), thumbImage);
-                }
-            } catch (Exception e) {
-                // ignore malformed uri, then attempt to extract resource ID.
-            }
-        }
-
-        String imageUrl = data.hasKey("imageUrl") ? data.getString("imageUrl") : null;
-        String type = data.hasKey("type") ? data.getString("type") : null;
-        if (type == null) {
-            callback.invoke(INVALID_ARGUMENT);
-        }
-
-        if (thumbUri == null) { // 无缩略图
-            this._share(scene, data, null, callback);
-        } else { // 有缩略图
-            if ("imageUrl".equals(type) || "imageResource".equals(type) || "imageFile".equals(type)) { // 分享图片
-                if (imageUrl == null) {
-                    callback.invoke(INVALID_ARGUMENT);
-                } else {
-                    final Uri finalThumbUri = thumbUri;
-                    getExtension(imageUrl, new ExtensionCallback() {
-                        @Override
-                        public void invoke(@Nullable String extension) {
-                            getImage(finalThumbUri, new ResizeOptions(100, 100), new ImageCallback() {
-                                @Override
-                                public void invoke(@Nullable Bitmap bitmap) {
-                                    _share(scene, data, bitmap, callback);
-                                }
-                            });
-                        }
-                    });
-                }
-            } else { // 分享非图片
-                _share(scene, data, null, callback);
-            }
-        }
-    }
-
-    private void _share(final int scene, final ReadableMap data, final Bitmap thumbImage, final Callback callback) {
-        if (!data.hasKey("type")) {
-            callback.invoke(INVALID_ARGUMENT);
-            return;
-        }
-        String type = data.getString("type");
-
-        WXMediaMessage.IMediaObject mediaObject = null;
-        if (type.equals("news")) {
-            mediaObject = _jsonToWebpageMedia(data);
-        } else if (type.equals("text")) {
-            mediaObject = _jsonToTextMedia(data);
-        } else if (type.equals("imageUrl") || type.equals("imageResource")) {
-            __jsonToImageUrlMedia(data, new MediaObjectCallback() {
-                @Override
-                public void invoke(@Nullable WXMediaMessage.IMediaObject mediaObject) {
-                    if (mediaObject == null) {
-                        callback.invoke(INVALID_ARGUMENT);
-                    } else {
-                        WeChatModule.this._share(scene, data, thumbImage, mediaObject, callback);
-                    }
-                }
-            });
-            return;
-        } else if (type.equals("imageFile")) {
-            __jsonToImageFileMedia(data, new MediaObjectCallback() {
-                @Override
-                public void invoke(@Nullable WXMediaMessage.IMediaObject mediaObject) {
-                    if (mediaObject == null) {
-                        callback.invoke(INVALID_ARGUMENT);
-                    } else {
-                        WeChatModule.this._share(scene, data, thumbImage, mediaObject, callback);
-                    }
-                }
-            });
-            return;
-        } else if (type.equals("video")) {
-            mediaObject = __jsonToVideoMedia(data);
-        } else if (type.equals("audio")) {
-            mediaObject = __jsonToMusicMedia(data);
-        } else if (type.equals("file")) {
-            mediaObject = __jsonToFileMedia(data);
-        }
-
-        if (mediaObject == null) {
-            callback.invoke(INVALID_ARGUMENT);
-        } else {
-            _share(scene, data, thumbImage, mediaObject, callback);
-        }
-    }
-
-    private void _share(int scene, ReadableMap data, Bitmap thumbImage, WXMediaMessage.IMediaObject mediaObject, Callback callback) {
-
-        WXMediaMessage message = new WXMediaMessage();
-        message.mediaObject = mediaObject;
-        if (thumbImage != null) {
-            message.setThumbImage(thumbImage);
-        }
-
-        if (data.hasKey("title")) {
-            message.title = data.getString("title");
-        }
-        if (data.hasKey("description")) {
-            message.description = data.getString("description");
-        }
-        if (data.hasKey("mediaTagName")) {
-            message.mediaTagName = data.getString("mediaTagName");
-        }
-        if (data.hasKey("messageAction")) {
-            message.messageAction = data.getString("messageAction");
-        }
-        if (data.hasKey("messageExt")) {
-            message.messageExt = data.getString("messageExt");
-        }
-
-        SendMessageToWX.Req req = new SendMessageToWX.Req();
-        req.message = message;
-        req.scene = scene;
-        req.transaction = UUID.randomUUID().toString();
-        callback.invoke(null, api.sendReq(req));
-    }
-
-    private WXTextObject _jsonToTextMedia(ReadableMap data) {
-        if (!data.hasKey("description")) {
-            return null;
-        }
-
-        WXTextObject ret = new WXTextObject();
-        ret.text = data.getString("description");
-        return ret;
-    }
-
-    private WXWebpageObject _jsonToWebpageMedia(ReadableMap data) {
-        if (!data.hasKey("webpageUrl")) {
-            return null;
-        }
-
-        WXWebpageObject ret = new WXWebpageObject();
-        ret.webpageUrl = data.getString("webpageUrl");
-        if (data.hasKey("extInfo")) {
-            ret.extInfo = data.getString("extInfo");
-        }
-        return ret;
-    }
-
-    private void __jsonToImageMedia(String imageUrl, final MediaObjectCallback callback) {
-        Uri imageUri;
         try {
-            imageUri = Uri.parse(imageUrl);
-            // Verify scheme is set, so that relative uri (used by static resources) are not handled.
-            if (imageUri.getScheme() == null) {
-                imageUri = getResourceDrawableUri(getReactApplicationContext(), imageUrl);
-            }
-        } catch (Exception e) {
-            imageUri = null;
+            callback.invoke(null, mWxPresenter.rnPay(data));
+        } catch (NotRegisterException | InvokeException e) {
+            callback.invoke(e.getMessage());
         }
-
-        if (imageUri == null) {
-            callback.invoke(null);
-            return;
-        }
-
-        final Uri finalImageUri = imageUri;
-        getExtension(imageUrl, new ExtensionCallback() {
-            @Override
-            public void invoke(String extension) {
-                if (extension != null && extension.toLowerCase().equals("gif")) {
-                    getImageData(finalImageUri, null, new ImageDataCallback() {
-                        @Override
-                        public void invoke(@Nullable byte[] bytes) {
-                            callback.invoke(bytes == null ? null : new WXEmojiObject(bytes));
-                        }
-                    });
-                } else {
-                    getImage(finalImageUri, new ResizeOptions(7000, 7000), new ImageCallback() {
-                        @Override
-                        public void invoke(@Nullable Bitmap bitmap) {
-                            callback.invoke(bitmap == null ? null : new WXImageObject(bitmap));
-                        }
-                    });
-                }
-            }
-        });
-
     }
 
-    private void __jsonToImageUrlMedia(ReadableMap data, MediaObjectCallback callback) {
-        if (!data.hasKey("imageUrl")) {
-            callback.invoke(null);
-            return;
-        }
-        String imageUrl = data.getString("imageUrl");
-        __jsonToImageMedia(imageUrl, callback);
-    }
+    private DataBus.OnBusResult<BaseResp> onWxResultOB = new DataBus.OnBusResult<BaseResp>() {
+        @Override
+        public void onResult(BaseResp baseResp) {
+            WritableMap map = Arguments.createMap();
+            map.putInt("errCode", baseResp.errCode);
+            map.putString("errStr", baseResp.errStr);
+            map.putString("openId", baseResp.openId);
+            map.putString("transaction", baseResp.transaction);
 
-    private void __jsonToImageFileMedia(ReadableMap data, MediaObjectCallback callback) {
-        if (!data.hasKey("imageUrl")) {
-            callback.invoke(null);
-            return;
-        }
-        String imageUrl = data.getString("imageUrl");
-        if (!imageUrl.toLowerCase().startsWith("file://")) {
-            imageUrl = "file://" + imageUrl;
-        }
-        __jsonToImageMedia(imageUrl, callback);
-    }
+            if (baseResp instanceof SendAuth.Resp) {
+                SendAuth.Resp resp = (SendAuth.Resp) (baseResp);
 
-    private WXMusicObject __jsonToMusicMedia(ReadableMap data) {
-        if (!data.hasKey("musicUrl")) {
-            return null;
-        }
-        WXMusicObject ret = new WXMusicObject();
-        ret.musicUrl = data.getString("musicUrl");
-        return ret;
-    }
-
-    private WXVideoObject __jsonToVideoMedia(ReadableMap data) {
-        if (!data.hasKey("videoUrl")) {
-            return null;
-        }
-        WXVideoObject ret = new WXVideoObject();
-        ret.videoUrl = data.getString("videoUrl");
-        return ret;
-    }
-
-    private WXFileObject __jsonToFileMedia(ReadableMap data) {
-        if (!data.hasKey("filePath")) {
-            return null;
-        }
-        return new WXFileObject(data.getString("filePath"));
-    }
-
-    private void getImage(Uri uri, ResizeOptions resizeOptions, final ImageCallback imageCallback) {
-        BaseBitmapDataSubscriber dataSubscriber = new BaseBitmapDataSubscriber() {
-            @Override
-            protected void onNewResultImpl(Bitmap bitmap) {
-                if (bitmap != null) {
-                    bitmap = bitmap.copy(bitmap.getConfig(), true);
-                }
-                imageCallback.invoke(bitmap);
+                map.putString("type", "SendAuth.Resp");
+                map.putString("code", resp.code);
+                map.putString("state", resp.state);
+                map.putString("url", resp.url);
+                map.putString("lang", resp.lang);
+                map.putString("country", resp.country);
+            } else if (baseResp instanceof SendMessageToWX.Resp) {
+                SendMessageToWX.Resp resp = (SendMessageToWX.Resp) (baseResp);
+                map.putString("type", "SendMessageToWX.Resp");
+            } else if (baseResp instanceof PayResp) {
+                PayResp resp = (PayResp) (baseResp);
+                map.putString("type", "PayReq.Resp");
+                map.putString("returnKey", resp.returnKey);
             }
 
-            @Override
-            protected void onFailureImpl(DataSource<CloseableReference<CloseableImage>> dataSource) {
-                imageCallback.invoke(null);
-            }
-        };
-
-        ImageRequestBuilder builder = ImageRequestBuilder.newBuilderWithSource(uri);
-        if (resizeOptions != null) {
-            builder = builder.setResizeOptions(resizeOptions);
+            getReactApplicationContext()
+                    .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                    .emit("WeChat_Resp", map);
         }
-        ImageRequest imageRequest = builder.build();
-
-        ImagePipeline imagePipeline = Fresco.getImagePipeline();
-        DataSource<CloseableReference<CloseableImage>> dataSource = imagePipeline.fetchDecodedImage(imageRequest, null);
-        dataSource.subscribe(dataSubscriber, UiThreadImmediateExecutorService.getInstance());
-    }
-
-    private void getImageData(Uri uri, ResizeOptions resizeOptions, final ImageDataCallback imageCallback) {
-        DataSubscriber<CloseableReference<PooledByteBuffer>> dataSubscriber =
-                new BaseDataSubscriber<CloseableReference<PooledByteBuffer>>() {
-
-                    @Override
-                    protected void onNewResultImpl(DataSource<CloseableReference<PooledByteBuffer>> dataSource) {
-                        // isFinished must be obtained before image, otherwise we might set intermediate result
-                        // as final image.
-                        boolean isFinished = dataSource.isFinished();
-                        CloseableReference<PooledByteBuffer> image = dataSource.getResult();
-                        if (image != null) {
-                            Preconditions.checkState(CloseableReference.isValid(image));
-                            PooledByteBuffer result = image.get();
-                            InputStream inputStream = new PooledByteBufferInputStream(result);
-                            byte[] bytes = null;
-                            try {
-                                bytes = getBytes(inputStream);
-                            } catch (IOException e) {
-                                e.printStackTrace();
-                            } finally {
-                                imageCallback.invoke(bytes);
-                                Closeables.closeQuietly(inputStream);
-                            }
-                        } else if (isFinished) {
-                            imageCallback.invoke(null);
-                        }
-                        dataSource.close();
-                    }
-
-                    @Override
-                    protected void onFailureImpl(DataSource<CloseableReference<PooledByteBuffer>> dataSource) {
-                        imageCallback.invoke(null);
-                    }
-                };
-
-        ImageRequestBuilder builder = ImageRequestBuilder.newBuilderWithSource(uri);
-        if (resizeOptions != null) {
-            builder = builder.setResizeOptions(resizeOptions);
-        }
-        ImageRequest imageRequest = builder.build();
-
-        ImagePipeline imagePipeline = Fresco.getImagePipeline();
-        DataSource<CloseableReference<PooledByteBuffer>> dataSource = imagePipeline.fetchEncodedImage(imageRequest, getReactApplicationContext());
-        dataSource.subscribe(dataSubscriber, UiThreadImmediateExecutorService.getInstance());
-    }
-
-    private byte[] getBytes(InputStream inputStream) throws IOException {
-        ByteArrayOutputStream byteBuffer = new ByteArrayOutputStream();
-        int bufferSize = 1024;
-        byte[] buffer = new byte[bufferSize];
-
-        int len = 0;
-        while ((len = inputStream.read(buffer)) != -1) {
-            byteBuffer.write(buffer, 0, len);
-        }
-        return byteBuffer.toByteArray();
-    }
-
-    private static Uri getResourceDrawableUri(Context context, String name) {
-        if (name == null || name.isEmpty()) {
-            return null;
-        }
-        name = name.toLowerCase().replace("-", "_");
-        int resId = context.getResources().getIdentifier(
-                name,
-                "drawable",
-                context.getPackageName());
-
-        if (resId == 0) {
-            return null;
-        } else {
-            return new Uri.Builder()
-                    .scheme(UriUtil.LOCAL_RESOURCE_SCHEME)
-                    .path(String.valueOf(resId))
-                    .build();
-        }
-    }
-
-    /**
-     * 获取链接指向文件后缀
-     *
-     * @param url
-     * @return
-     */
-    private void getExtension(final String url, final ExtensionCallback callback) {
-        if (url == null) {
-            callback.invoke(null);
-            return;
-        }
-        String temp = extensionMap.get(url);
-        if (temp != null) {
-            callback.invoke(temp);
-            return;
-        }
-        Uri uri = Uri.parse(url);
-        String scheme = uri.getScheme();
-        if (scheme == null || scheme.equals("file")) {
-            String path = uri.getPath();
-            int index = path.lastIndexOf(".");
-            String extension;
-            if (index > 0) {
-                extension = path.substring(index + 1);
-            } else {
-                extension = null;
-            }
-            if (extension != null) {
-                extensionMap.put(url, extension);
-            }
-            callback.invoke(extension);
-        } else {
-            Request request = new Request.Builder().url(url).head().build();
-            OkHttpClient client = OkHttpClientProvider.getOkHttpClient();
-            client.newCall(request).enqueue(new okhttp3.Callback() {
-                @Override
-                public void onFailure(Call call, IOException e) {
-                    callback.invoke(null);
-                }
-
-                @Override
-                public void onResponse(Call call, Response response) throws IOException {
-                    String contentType = response.body().contentType().toString();
-                    String extension = MimeTypeMap.getSingleton().getExtensionFromMimeType(contentType);
-                    if (extension != null) {
-                        extensionMap.put(url, extension);
-                    }
-                    callback.invoke(extension);
-                }
-            });
-        }
-
-    }
-
-    // TODO: 实现sendRequest、sendSuccessResponse、sendErrorCommonResponse、sendErrorUserCancelResponse
-
-    @Override
-    public void onReq(BaseReq baseReq) {
-
-    }
-
-    @Override
-    public void onResp(BaseResp baseResp) {
-        WritableMap map = Arguments.createMap();
-        map.putInt("errCode", baseResp.errCode);
-        map.putString("errStr", baseResp.errStr);
-        map.putString("openId", baseResp.openId);
-        map.putString("transaction", baseResp.transaction);
-
-        if (baseResp instanceof SendAuth.Resp) {
-            SendAuth.Resp resp = (SendAuth.Resp) (baseResp);
-
-            map.putString("type", "SendAuth.Resp");
-            map.putString("code", resp.code);
-            map.putString("state", resp.state);
-            map.putString("url", resp.url);
-            map.putString("lang", resp.lang);
-            map.putString("country", resp.country);
-        } else if (baseResp instanceof SendMessageToWX.Resp) {
-            SendMessageToWX.Resp resp = (SendMessageToWX.Resp) (baseResp);
-            map.putString("type", "SendMessageToWX.Resp");
-        } else if (baseResp instanceof PayResp) {
-            PayResp resp = (PayResp) (baseResp);
-            map.putString("type", "PayReq.Resp");
-            map.putString("returnKey", resp.returnKey);
-        }
-
-        this.getReactApplicationContext()
-                .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
-                .emit("WeChat_Resp", map);
-    }
-
-    private interface ImageCallback {
-        void invoke(@Nullable Bitmap bitmap);
-    }
-
-    private interface ImageDataCallback {
-        void invoke(@Nullable byte[] bytes);
-    }
-
-    private interface MediaObjectCallback {
-        void invoke(@Nullable WXMediaMessage.IMediaObject mediaObject);
-    }
-
-    private interface ExtensionCallback {
-        void invoke(@Nullable String extension);
-    }
+    };
 
 }
