+++ android/src/main/java/com/theweflex/react/exception/InvalidArgumentException.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2018. chenqiang Inc. All rights reserved.
+ */
+
+package com.theweflex.react.exception;
+
+/**
+ * {描述信息}
+ *
+ * @author chenqiang
+ * @date 2018/8/9
+ */
+public class InvalidArgumentException extends Exception {
+    public InvalidArgumentException() {
+        super("invalid argument.");
+    }
+}
