+++ android/src/main/java/com/theweflex/react/exception/NotRegisterException.java
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2018. chenqiang Inc. All rights reserved.
+ */
+
+package com.theweflex.react.exception;
+
+/**
+ * {描述信息}
+ *
+ * @author chenqiang
+ * @date 2018/8/9
+ */
+public class NotRegisterException extends Exception {
+
+    public NotRegisterException() {
+        super("registerApp required.");
+    }
+}
